wix-automation/requirements.txt
requests==2.32.0
pandas==2.2.0
python-dotenv==1.0.1
PyGithub==2.1.1
imap-tools==1.7.0
wix-analytics-python==1.0.0
